fix(sql): preselect ancestor update ids once
